+++ src/Segment.cc
@@ -411,8 +411,7 @@ Segment::getAppendedLength(Certificate* certificate) const
         certificate->segmentLength = head;
         Crc32C certificateChecksum = checksum;
         certificateChecksum.update(
-            certificate, static_cast<unsigned>
-            (sizeof(*certificate) - sizeof(certificate->checksum)));
+            certificate, sizeof(*certificate) - sizeof(certificate->checksum));
         certificate->checksum = certificateChecksum.getResult();
     }
     return head;
@@ -525,8 +524,8 @@ Segment::checkMetadataIntegrity(const Certificate& certificate)
         return false;
     }
 
-    currentChecksum.update(&certificate, static_cast<unsigned>
-                           (sizeof(certificate)-sizeof(certificate.checksum)));
+    currentChecksum.update(&certificate,
+                           sizeof(certificate)-sizeof(certificate.checksum));
 
     if (certificate.checksum != currentChecksum.getResult()) {
         LOG(WARNING, "segment corrupt: bad checksum (expected 0x%08x, "
